requirements.txt
swig==4.3.0
box2d-py==2.3.5
contourpy==1.3.1
fonttools==4.57.0
gymnasium==1.1.1
kiwisolver==1.4.8
matplotlib==3.10.1
numpy==2.2.4
pillow==11.1.0
pygame==2.6.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
six==1.17.0

